refactor relay calculations to its own file
